fix(sse): sse events for multi-cluster
