+++ backend/handlers/base/base.go
@@ -35,68 +35,38 @@ type BaseHandler struct {
 }
 
 func (h *BaseHandler) GetList(c echo.Context) error {
-	go func() {
-		<-c.Request().Context().Done()
-	}()
-
-	h.Container.SSE().ServeHTTP(h.Kind, c.Response(), c.Request())
-	return nil
-}
-
-func (h *BaseHandler) processList(resourceName string) func() {
-	return func() {
-		items := h.Informer.GetStore().List()
-		data := h.marshalListData(items, resourceName)
-		h.Container.SSE().Publish(h.Kind, &sse.Event{
-			Data: data,
-		})
-	}
-}
-
-func (h *BaseHandler) GetDetails(c echo.Context) error {
-	go func() {
-		<-c.Request().Context().Done()
-	}()
-	streamID, _, _, err := h.getStreamIDAndItem(c.QueryParam("namespace"), c.Param("name"))
-	if err != nil {
-		return c.String(http.StatusBadRequest, err.Error())
-	}
-	h.Event.AddEvent(streamID, h.processDetails(c.QueryParam("namespace"), c.Param("name")))
+	streamID := fmt.Sprintf("%s-%s-%s", h.QueryConfig, h.QueryCluster, h.Kind)
 	h.Container.SSE().ServeHTTP(streamID, c.Response(), c.Request())
 	return nil
 }
 
-func (h *BaseHandler) processDetails(namespace, name string) func() {
-	return func() {
-		streamID, item, exists, _ := h.getStreamIDAndItem(namespace, name)
-		data := h.marshalDetailData(item, exists)
-		h.Container.SSE().Publish(streamID, &sse.Event{
-			Data: data,
-		})
-	}
-}
-
-func (h *BaseHandler) GetYaml(c echo.Context) error {
-	go func() {
-		<-c.Request().Context().Done()
-	}()
-	streamID, _, _, err := h.getStreamIDAndItem(c.QueryParam("namespace"), c.Param("name"))
+func (h *BaseHandler) GetDetails(c echo.Context) error {
+	streamID, item, exists, err := h.getStreamIDAndItem(c.QueryParam("namespace"), c.Param("name"))
 	if err != nil {
 		return c.String(http.StatusBadRequest, err.Error())
 	}
-	h.Event.AddEvent(streamID, h.processYAML(c.QueryParam("namespace"), c.Param("name")))
-	h.Container.SSE().ServeHTTP(fmt.Sprintf("%s-yaml", streamID), c.Response(), c.Request())
+	go func() {
+		h.Container.SSE().Publish(streamID, &sse.Event{
+			Data: h.marshalDetailData(item, exists),
+		})
+	}()
+	h.Container.SSE().ServeHTTP(streamID, c.Response(), c.Request())
 	return nil
 }
 
-func (h *BaseHandler) processYAML(namespace, name string) func() {
-	return func() {
-		streamID, item, exists, _ := h.getStreamIDAndItem(namespace, name)
-		data := h.marshalYAML(item, exists)
-		h.Container.SSE().Publish(fmt.Sprintf("%s-yaml", streamID), &sse.Event{
-			Data: data,
-		})
+func (h *BaseHandler) GetYaml(c echo.Context) error {
+	streamID, item, exists, err := h.getStreamIDAndItem(c.QueryParam("namespace"), c.Param("name"))
+	if err != nil {
+		return c.String(http.StatusBadRequest, err.Error())
 	}
+	go func() {
+		h.Container.SSE().Publish(fmt.Sprintf("%s-yaml", streamID), &sse.Event{
+			Data: h.marshalYAML(item, exists),
+		})
+	}()
+
+	h.Container.SSE().ServeHTTP(fmt.Sprintf("%s-yaml", streamID), c.Response(), c.Request())
+	return nil
 }
 
 func (h *BaseHandler) GetEvents(c echo.Context) error {
