+++ backend/handlers/base/informers.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"github.com/charmbracelet/log"
 	"github.com/labstack/echo/v4"
+	"github.com/r3labs/sse/v2"
 	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/client-go/tools/cache"
 	"strings"
@@ -20,14 +21,14 @@ func ResourceEventHandler[T Resource](handler *BaseHandler, additionalEvents ...
 	handleEvent := func(obj any) {
 		resource := obj.(T)
 		// GetList
-		go handler.Event.AddEvent(handler.Kind, handler.processList(resource.GetName()))
+		go handler.Event.AddEvent(handler.Kind, handler.processListEvents(resource.GetName()))
 
 		streamName := fmt.Sprintf("%s-%s", resource.GetNamespace(), resource.GetName())
 		// GetDetails
-		go handler.Event.AddEvent(streamName, handler.processDetails(resource.GetNamespace(), resource.GetName()))
+		go handler.Event.AddEvent(streamName, handler.processDetailsEvents(resource.GetNamespace(), resource.GetName()))
 
 		// GetYAML
-		go handler.Event.AddEvent(streamName+"-yaml", handler.processYAML(resource.GetNamespace(), resource.GetName()))
+		go handler.Event.AddEvent(streamName+"-yaml", handler.processYAMLEvents(resource.GetNamespace(), resource.GetName()))
 
 		for _, event := range additionalEvents {
 			go handler.Event.AddEvent("deployments", event)
@@ -69,7 +70,7 @@ func (h *BaseHandler) WaitForSync(c echo.Context) {
 	err := wait.PollUntilContextCancel(c.Request().Context(), 100*time.Millisecond, true, func(context.Context) (done bool, err error) {
 		hasSynced := h.Informer.HasSynced()
 		if hasSynced {
-			h.Event.AddEvent(h.Kind, h.processList(""))
+			h.Event.AddEvent(h.Kind, h.processListEvents(""))
 		}
 		return hasSynced, nil
 	})
@@ -94,4 +95,35 @@ func (h *BaseHandler) isNamespaceResource(r string) bool {
 	}
 
 	return true
+}
+
+func (h *BaseHandler) processListEvents(resourceName string) func() {
+	return func() {
+		items := h.Informer.GetStore().List()
+		data := h.marshalListData(items, resourceName)
+		streamID := fmt.Sprintf("%s-%s-%s", h.QueryConfig, h.QueryCluster, h.Kind)
+		h.Container.SSE().Publish(streamID, &sse.Event{
+			Data: data,
+		})
+	}
+}
+
+func (h *BaseHandler) processDetailsEvents(namespace, name string) func() {
+	return func() {
+		streamID, item, exists, _ := h.getStreamIDAndItem(namespace, name)
+		data := h.marshalDetailData(item, exists)
+		h.Container.SSE().Publish(streamID, &sse.Event{
+			Data: data,
+		})
+	}
+}
+
+func (h *BaseHandler) processYAMLEvents(namespace, name string) func() {
+	return func() {
+		streamID, item, exists, _ := h.getStreamIDAndItem(namespace, name)
+		data := h.marshalYAML(item, exists)
+		h.Container.SSE().Publish(fmt.Sprintf("%s-yaml", streamID), &sse.Event{
+			Data: data,
+		})
+	}
 }
